Added EKS support with a goal to remove ECS
This launches an EKS cluster and sets up a fargate
profile. Since we are using a lot of shared resources
I elected to setup the cluster with terraform rather than
eksctl.

!!!WARNING!!! - rant ahead

<rant>
ECS with fargate makes it unnecessarily hard to run one
off tasks.  The postgraphile container on ECS cluster
was on a crashloop backoff but I wanted to simply
launch an one off task to run some migrations.
It was a long yak shaving session that let me down to
fargatecli (https://github.com/awslabs/fargatecli).

I would give AWS an 'A' for effort but it does not
even get close to kubectl run. Sometimes even the
best written software needs a shell to run an one-off
task. And the reality is that most software is not
well written. The suggestion to get a shell into
your task is to run a SSH enabled container.

Seriously?!!! I am throwing in the towel on
ECS and moving to EKS for all my further deployments.

And don't even get me started with the AWS session manager
shenanigans. SSH is a protocol that is simple and that
works. Instead of working with certificates and
revocation lists and keeping private keys private,
they want to peddle their KMS. Kubernetes does not
support ssh into its running containers but it does
support port-forwarding and socat rather elegantly
out of the box. I am all for choices with our deployment
tools but I'd rather stick with kubernetes than some
weird AWS specific way of doing things.
</rant>