[#314] - Setup Dockerfile for mealplanner-ui and admin-ui
We had to make a few changes to the routing in mealplanner-ui. Since we
do not have a backend component to handle routes, we are changing this to
a hash router instead. Dockerfiles are designed to mimic production. Production's
Dockerfiles should be simpler without needing nginx.conf patching. We can handle that
as a part of an ingress resource. For mealplanner-ui we need to mount graphql
endpoint on `/graphql` and for admin ui, we need it mounted on `/api/graphql`.

Co-authored-by: Vagmi Mudumbai <vagmi.mudumbai@gmail.com>
Co-authored-by: Shanthi Shanmugam <shanthi.shanmugam@gmail.com>